Rename the new adapter to a more fitting name
